move cache tests to Bedrock test file
Replace the module-level pytest.skip with per-test skipif markers so
that unit tests (which don't need credentials) can run alongside the
live-API tests.
