+++ tests/test_provider_anthropic.py
@@ -10,7 +10,7 @@ from chatlas import (
     tool_web_fetch,
     tool_web_search,
 )
-from chatlas._provider_anthropic import AnthropicBedrockProvider, AnthropicProvider
+from chatlas._provider_anthropic import AnthropicProvider
 from pydantic import BaseModel, Field
 
 from .conftest import (
@@ -303,28 +303,3 @@ def test_anthropic_nested_data_model_extraction():
         assert 0.0 <= classification.score <= 1.0, (
             f"Score {classification.score} should be between 0 and 1"
         )
-
-
-def _make_bedrock_provider(**kwargs):
-    defaults = dict(
-        model="us.anthropic.claude-sonnet-4-6",
-        aws_secret_key="fake",
-        aws_access_key="fake",
-        aws_region="us-east-1",
-        aws_profile=None,
-        aws_session_token=None,
-        base_url=None,
-    )
-    return AnthropicBedrockProvider(**{**defaults, **kwargs})
-
-
-class TestBedrockCacheDefault:
-    def test_default_enables_caching(self):
-        provider = _make_bedrock_provider()
-        assert provider._cache == "5m"
-        assert provider._cache_control() == {"type": "ephemeral", "ttl": "5m"}
-
-    def test_none_disables_caching(self):
-        provider = _make_bedrock_provider(cache="none")
-        assert provider._cache == "none"
-        assert provider._cache_control() is None
